finishing add button

+++ Laravel/Laravel/app/Http/Controllers/UserinfoController.php
@@ -3,29 +3,47 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+// use Illuminate\Http\Response;
 use App\Models\Userinfo;
 
 class UserinfoController extends Controller
 {
-    public function index()
+    public function index(Request $req)
     {
-        return Userinfo::all();  
+        if ($req->is('api/userinfo/index'))
+        {
+            return Userinfo::all();
+        }
     }
     
-    public function delete($id)
+    public function delete(Request $req, $id)
     {
-        Userinfo::find($id)->delete();
+        if ($req->is('api/userinfo/delete/' + $id))
+        {
+            Userinfo::find($id)->delete();
+            return response("OK",200);
+        }
+        
     }
 
     public function add(Request $req)
+    {       
+        if ($req->is('api/userinfo/add'))
+        {
+            $data = [
+                "name" => $req->input('name'),
+                "gender" => $req->input('gender'),
+                "age" => $req->input('age'),
+                "number" => $req->input('number'),
+                "email" => $req->input('email'),
+            ];
+            Userinfo::create($data);
+            return response($data);
+        }
+    }
+
+    public function testt(Request $req)
     {
-        $data = [
-            "name" => $req->name,
-            "gender" => $req->gender,
-            "age" => $req->age,
-            "number" => $req->number,
-            "email" => $req->email,
-        ];
-        Userinfo::create($data);        
+        
     }
 }
